requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.12.4
openai
langfuse==3.00.0
jinja2==3.1.2
google-genai
websockets>=13.0,<15.0
openinference-instrumentation-google-genai==0.1.0
supabase==2.24.0
supabase-auth==2.24.0
python-dotenv
